moved getShowConnectAccountToast()
